Sync Fork from Upstream Repo

+++ src/controller/p2p/preheat/controller.go
@@ -2,9 +2,9 @@ package preheat
 
 import (
 	"context"
-	"encoding/json"
 	"time"
 
+	"github.com/goharbor/harbor/src/jobservice/job"
 	"github.com/goharbor/harbor/src/lib/errors"
 	"github.com/goharbor/harbor/src/lib/q"
 	"github.com/goharbor/harbor/src/pkg/p2p/preheat/instance"
@@ -13,6 +13,7 @@ import (
 	"github.com/goharbor/harbor/src/pkg/p2p/preheat/policy"
 	"github.com/goharbor/harbor/src/pkg/p2p/preheat/provider"
 	"github.com/goharbor/harbor/src/pkg/scheduler"
+	"github.com/goharbor/harbor/src/pkg/task"
 )
 
 const (
@@ -116,6 +117,8 @@ type Controller interface {
 	ListPoliciesByProject(ctx context.Context, project int64, query *q.Query) ([]*policyModels.Schema, error)
 	// CheckHealth checks the instance health, for test connection
 	CheckHealth(ctx context.Context, instance *providerModels.Instance) error
+	// DeletePoliciesOfProject delete all policies under one project
+	DeletePoliciesOfProject(ctx context.Context, project int64) error
 }
 
 var _ Controller = (*controller)(nil)
@@ -126,16 +129,18 @@ type controller struct {
 	// For instance
 	iManager instance.Manager
 	// For policy
-	pManager  policy.Manager
-	scheduler scheduler.Scheduler
+	pManager     policy.Manager
+	scheduler    scheduler.Scheduler
+	executionMgr task.ExecutionManager
 }
 
 // NewController is constructor of controller
 func NewController() Controller {
 	return &controller{
-		iManager:  instance.Mgr,
-		pManager:  policy.Mgr,
-		scheduler: scheduler.Sched,
+		iManager:     instance.Mgr,
+		pManager:     policy.Mgr,
+		scheduler:    scheduler.Sched,
+		executionMgr: task.NewExecutionManager(),
 	}
 }
 
@@ -184,11 +189,59 @@ func (c *controller) CreateInstance(ctx context.Context, instance *providerModel
 
 // DeleteInstance implements @Controller.Delete
 func (c *controller) DeleteInstance(ctx context.Context, id int64) error {
+	ins, err := c.GetInstance(ctx, id)
+	if err != nil {
+		return err
+	}
+	// delete instance should check the instance whether be used by policies
+	policies, err := c.ListPolicies(ctx, &q.Query{
+		Keywords: map[string]interface{}{
+			"provider_id": id,
+		},
+	})
+	if err != nil {
+		return err
+	}
+
+	if len(policies) > 0 {
+		return errors.New(nil).
+			WithCode(errors.PreconditionCode).
+			WithMessage("Provider [%s] cannot be deleted as some preheat policies are using it", ins.Name)
+	}
+
 	return c.iManager.Delete(ctx, id)
 }
 
 // UpdateInstance implements @Controller.Update
 func (c *controller) UpdateInstance(ctx context.Context, instance *providerModels.Instance, properties ...string) error {
+	oldIns, err := c.GetInstance(ctx, instance.ID)
+	if err != nil {
+		return err
+	}
+
+	if !instance.Enabled {
+		// update instance should check the instance whether be used by policies
+		policies, err := c.ListPolicies(ctx, &q.Query{
+			Keywords: map[string]interface{}{
+				"provider_id": instance.ID,
+			},
+		})
+		if err != nil {
+			return err
+		}
+
+		if len(policies) > 0 {
+			return errors.New(nil).
+				WithCode(errors.PreconditionCode).
+				WithMessage("Provider [%s] cannot be disabled as some preheat policies are using it", oldIns.Name)
+		}
+	}
+
+	// vendor type does not support change
+	if oldIns.Vendor != instance.Vendor {
+		return errors.Errorf("provider [%s] vendor cannot be changed", oldIns.Name)
+	}
+
 	return c.iManager.Update(ctx, instance, properties...)
 }
 
@@ -223,7 +276,7 @@ func (c *controller) CreatePolicy(ctx context.Context, schema *policyModels.Sche
 	schema.UpdatedTime = now
 
 	// Get full model of policy schema
-	_, err = policy.ParsePolicy(schema)
+	err = schema.Decode()
 	if err != nil {
 		return 0, err
 	}
@@ -239,17 +292,17 @@ func (c *controller) CreatePolicy(ctx context.Context, schema *policyModels.Sche
 		schema.Trigger.Type == policyModels.TriggerTypeScheduled &&
 		len(schema.Trigger.Settings.Cron) > 0 {
 		// schedule and update policy
-		schema.Trigger.Settings.JobID, err = c.scheduler.Schedule(ctx, schema.Trigger.Settings.Cron, SchedulerCallback, TriggerParam{PolicyID: id})
-		if err != nil {
+		if _, err = c.scheduler.Schedule(ctx, job.P2PPreheat, id, "", schema.Trigger.Settings.Cron,
+			SchedulerCallback, TriggerParam{PolicyID: id}); err != nil {
 			return 0, err
 		}
 
-		if err = decodeSchema(schema); err == nil {
+		if err = schema.Encode(); err == nil {
 			err = c.pManager.Update(ctx, schema, "trigger")
 		}
 
 		if err != nil {
-			if e := c.scheduler.UnSchedule(ctx, schema.Trigger.Settings.JobID); e != nil {
+			if e := c.scheduler.UnScheduleByVendor(ctx, job.P2PPreheat, id); e != nil {
 				return 0, errors.Wrap(e, err.Error())
 			}
 
@@ -288,18 +341,18 @@ func (c *controller) UpdatePolicy(ctx context.Context, schema *policyModels.Sche
 	}
 
 	// Get full model of updating policy
-	_, err = policy.ParsePolicy(schema)
+	err = schema.Decode()
 	if err != nil {
 		return err
 	}
 
 	var cron = schema.Trigger.Settings.Cron
-	var oldJobID = s0.Trigger.Settings.JobID
+	var oldCron = s0.Trigger.Settings.Cron
 	var needUn bool
 	var needSch bool
 
 	if s0.Trigger.Type != schema.Trigger.Type {
-		if s0.Trigger.Type == policyModels.TriggerTypeScheduled && oldJobID > 0 {
+		if s0.Trigger.Type == policyModels.TriggerTypeScheduled && len(oldCron) > 0 {
 			needUn = true
 		}
 		if schema.Trigger.Type == policyModels.TriggerTypeScheduled && len(cron) > 0 {
@@ -307,10 +360,9 @@ func (c *controller) UpdatePolicy(ctx context.Context, schema *policyModels.Sche
 		}
 	} else {
 		// not change trigger type
-		if schema.Trigger.Type == policyModels.TriggerTypeScheduled &&
-			s0.Trigger.Settings.Cron != cron {
+		if schema.Trigger.Type == policyModels.TriggerTypeScheduled && oldCron != cron {
 			// unschedule old
-			if oldJobID > 0 {
+			if len(oldCron) > 0 {
 				needUn = true
 			}
 			// schedule new
@@ -324,7 +376,7 @@ func (c *controller) UpdatePolicy(ctx context.Context, schema *policyModels.Sche
 
 	// unschedule old
 	if needUn {
-		err = c.scheduler.UnSchedule(ctx, oldJobID)
+		err = c.scheduler.UnScheduleByVendor(ctx, job.P2PPreheat, schema.ID)
 		if err != nil {
 			return err
 		}
@@ -332,22 +384,21 @@ func (c *controller) UpdatePolicy(ctx context.Context, schema *policyModels.Sche
 
 	// schedule new
 	if needSch {
-		jobid, err := c.scheduler.Schedule(ctx, cron, SchedulerCallback, TriggerParam{PolicyID: schema.ID})
-		if err != nil {
+		if _, err := c.scheduler.Schedule(ctx, job.P2PPreheat, schema.ID, "", cron, SchedulerCallback,
+			TriggerParam{PolicyID: schema.ID}); err != nil {
 			return err
-		}
-		schema.Trigger.Settings.JobID = jobid
-		if err := decodeSchema(schema); err != nil {
-			// Possible
-			// TODO: Refactor
-			return err // whether update or not has been not important as the jon reference will be lost
 		}
 	}
 
 	// Update timestamp
 	schema.UpdatedTime = time.Now()
 
-	return c.pManager.Update(ctx, schema, props...)
+	err = c.pManager.Update(ctx, schema, props...)
+	if (err != nil) && (needSch || needUn) {
+		return errors.Wrapf(err, "Update failed, but not rollback scheduler")
+	}
+
+	return err
 }
 
 // DeletePolicy deletes the policy by id.
@@ -356,14 +407,55 @@ func (c *controller) DeletePolicy(ctx context.Context, id int64) error {
 	if err != nil {
 		return err
 	}
-	if s.Trigger != nil && s.Trigger.Type == policyModels.TriggerTypeScheduled && s.Trigger.Settings.JobID > 0 {
-		err = c.scheduler.UnSchedule(ctx, s.Trigger.Settings.JobID)
+	if s.Trigger != nil && s.Trigger.Type == policyModels.TriggerTypeScheduled && len(s.Trigger.Settings.Cron) > 0 {
+		err = c.scheduler.UnScheduleByVendor(ctx, job.P2PPreheat, id)
 		if err != nil {
 			return err
 		}
 	}
 
+	if err = c.deleteExecs(ctx, id); err != nil {
+		return err
+	}
+
 	return c.pManager.Delete(ctx, id)
+}
+
+// DeletePoliciesOfProject deletes all the policy under project.
+func (c *controller) DeletePoliciesOfProject(ctx context.Context, project int64) error {
+	policies, err := c.ListPoliciesByProject(ctx, project, nil)
+	if err != nil {
+		return err
+	}
+
+	for _, p := range policies {
+		if err = c.DeletePolicy(ctx, p.ID); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// deleteExecs delete executions
+func (c *controller) deleteExecs(ctx context.Context, vendorID int64) error {
+	executions, err := c.executionMgr.List(ctx, &q.Query{
+		Keywords: map[string]interface{}{
+			"VendorType": job.P2PPreheat,
+			"VendorID":   vendorID,
+		},
+	})
+
+	if err != nil {
+		return err
+	}
+
+	for _, execution := range executions {
+		if err = c.executionMgr.Delete(ctx, execution.ID); err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // ListPolicies lists policies by query.
@@ -402,23 +494,6 @@ func (c *controller) CheckHealth(ctx context.Context, instance *providerModels.I
 	if h.Status != provider.DriverStatusHealthy {
 		return errors.Errorf("preheat provider instance %s-%s:%s is not healthy", instance.Vendor, instance.Name, instance.Endpoint)
 	}
-
-	return nil
-}
-
-// decodeSchema decodes the trigger object to JSON string
-func decodeSchema(schema *policyModels.Schema) error {
-	if schema.Trigger == nil {
-		// do nothing
-		return nil
-	}
-
-	b, err := json.Marshal(schema.Trigger)
-	if err != nil {
-		return err
-	}
-
-	schema.TriggerStr = string(b)
 
 	return nil
 }
